+++ helpers/controllers/lessonsController.ts
@@ -8,7 +8,7 @@ import type {
 import { lessons } from '../../graphql/queryResolvers/lessons'
 import prisma from '../../prisma'
 import { isAdminOrThrow } from '../isAdmin'
-import { validateLessonId } from '../validateLessonId'
+import { validateLessonId } from '../validation/validateLessonId'
 
 export const createLesson = async (
   _parent: void,
