+++ helpers/validation/validateLessonId.ts
@@ -1,4 +1,4 @@
-import prisma from '../prisma'
+import prisma from '../../prisma'
 
 export const validateLessonId = async (
   id: number,
